[feat] 정산 기본 구현 및 목록 조회 쿼리DSL # 57

+++ backend/src/main/java/com/synerge/order101/purchase/model/entity/PurchaseDetail.java
@@ -31,7 +31,7 @@ public class PurchaseDetail {
     private Product product;
 
     @Column
-    private Double orderQty;
+    private Integer orderQty;
 
     @Column
     private Double unitPrice;
